pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nblane"
version = "0.2.0"
description = "Human + Agent + Team co-evolution system"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "streamlit>=1.35",
    "openai>=1.30",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "mcp>=1.0",
]

[project.scripts]
nblane = "nblane.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
